refactor: starter 스크린의 컴포넌트 폴더 구조 컨벤션에 맞추어 변경
#2
