+++ src/components/feature/screens/Starter/LoginButton/index.tsx
@@ -1,0 +1,12 @@
+import PositionBottom from "@components/common/shared/layout/PositionBottom";
+import Button from "@components/common/shared/ui/Button";
+
+const LoginButton = () => {
+  return (
+    <PositionBottom>
+      <Button title="로그인 하러 가기" fullWidth />
+    </PositionBottom>
+  );
+};
+
+export default LoginButton;
